fix (kein Essen) problems with latest season in Berlin update website
